requirements.txt
adb-shell==0.4.4
aiohttp==3.11.2
bs4==0.0.2
dnspython==2.7.0
portscan==1.1
python-dateutil==2.9.0.post0
requests==2.32.3
tabulate==0.9.0
tzdata==2024.2
